Bump to moodle-cs ^3.3.1
Note that we have, also, to explicitly require:

phpcompatibility/php-compatibility:dev-develop#2fb82334

Because the one defined in moodle-cs requires to change
the global minimum-stability of the project or older
versions continue being used (moodle-cs 2.3.5 and
phpcompatibility 9.3.5).

By defining it like own requirement, it's allowed
to use the dev version without problem.

This requirement will be removed once phpcompatibility
releases a new (10.x) version.

Also, tidy unit tests to verify phpcompatibility sniffs
are working as expected for php 7.4 and up.