Force php-compatibility dev to be used
Since moodle-cs 3.3.0 we are using "phpcompatibility/php-compatibility": "dev-develop#2fb82334"
because the last version released (3.9.5) is from 2019 and missing lots
of stuff and, also, compatibility with php81 and up.

To be able to install those new moodle-cs versions, all products
installing it must explicitly require that php-compatibility version.

Once the make a new release, we'll be able to revert this.