Prevent plugins to check FileExpectedTags.LicenseTagInvalid
A new `license-regex` option has been added to the phpcs command.

It allows to specify the regex that will be used when inspecting
@license tags.

By default, an empty string is used for the option and that makes
the Sniff to stop checking. Whoever wants to check for anything
can use the new option to configure it.