Deprecate the phpdbg option for code coverage
This includes both the current default (when neither pcov or
xdebug are available) and the explicit use of --coverage-phpdbg

Since PHP 8.0, phpdbg is causing all sort of problems and is
not recommended. So it's time to deprecate it in order to
proceed to remove it completely for moodle-plugin-ci 5.0.0.

Removal ref.: https://github.com/moodlehq/moodle-plugin-ci/issues/289

Fixes #213